qml: remove symbol export and private header
it's a plugin. no symbol need to be exported
