+++ qml/QmlAV/Export.h
@@ -1,6 +1,6 @@
 /******************************************************************************
     QtAV:  Media play library based on Qt and FFmpeg
-    Copyright (C) 2013 Wang Bin <wbsecg1@gmail.com>
+    Copyright (C) 2013-2015 Wang Bin <wbsecg1@gmail.com>
 
 *   This file is part of QtAV
 
@@ -32,5 +32,4 @@
 #  define QMLAV_EXPORT Q_DECL_IMPORT //only for vc?
 #endif
 
-#define CONFIG_FBO_ITEM (QT_VERSION >= QT_VERSION_CHECK(5, 2, 0))
 #endif // QTAV_QML_EXPORT_H
